Fix: 스타일 컴포넌트 파일분리한 것 하나로 합침작업
